Reorganize folders following DDD
