+++ src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/variables.scss";
 
 .app-container {
   height: 100vh;
