+++ src/domains/login/pages/login-page/login-page.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "src/variables.scss";
 
 .login-page__container {
   display: flex;
