+++ src/domains/login/pages/login-page/login-page.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnDestroy } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
 import { Subscription } from 'rxjs';
-import { AccountService } from 'src/services/AccountService';
-import { LocalStorageService } from 'src/services/LocalStorageService';
-import { User } from 'src/utils/interfaces/User';
+import { AccountService } from 'src/domains/login/services/AccountService';
+import { LocalStorageService } from 'src/domains/shared/services/LocalStorageService';
+import { User } from 'src/domains/login/models/User';
 
 @Component({
   selector: 'app-login-page',
