+++ src/domains/payment/components/modal/modal.component.spec.ts
