+++ src/domains/payment/components/table/table.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "src/variables.scss";
 
 .table__container {
   border-radius: 8px;
