+++ src/domains/payment/pages/tasks-page/tasks-page.component.scss
@@ -1,5 +1,5 @@
-@import "../../variables.scss";
-@import "../../styles.scss";
+@import "src/variables.scss";
+@import "src/styles.scss";
 
 .tasks-page__container {
   font-family: 'Roboto', sans-serif;
