+++ src/domains/payment/pages/tasks-page/tasks-page.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from "@angular/core";
-import { NavbarModule } from "src/components/navbar/navbar.module";
+import { NavbarModule } from "src/domains/shared/components/navbar/navbar.module";
 import { TasksPageComponent } from "./tasks-page.component";
-import { TableComponentModule } from "src/components/table/table.component.module";
+import { TableComponentModule } from "src/domains/payment/components/table/table.component.module";
 import { BrowserModule } from "@angular/platform-browser";
-import { ModalComponentModule } from "src/shared/components/modal/modal.component.module";
 import { MatDialogModule } from "@angular/material/dialog";
+import { ModalComponentModule } from "../../components/modal/modal.component.module";
 
 @NgModule({
   declarations: [
