+++ src/domains/payment/utils/format.ts
