+++ src/domains/shared/components/navbar/navbar.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "src/variables.scss";
 
 .navbar {
   background-color: $dark-blue;
