+++ src/domains/shared/components/navbar/navbar.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
-import { AccountService } from 'src/services/AccountService';
-import { LocalStorageService } from 'src/services/LocalStorageService';
+import { AccountService } from 'src/domains/login/services/AccountService';
+import { LocalStorageService } from 'src/domains/shared/services/LocalStorageService';
 import { NavbarComponent } from './navbar.component';
 @NgModule({
   declarations: [
